+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Extended descriptions to the `SpaceCat` module commands
+
+### Changed
+- Revamp the `help` menu to use embeds for pretty formatting
+- `help` command without arguments now only shows modules to reduce clutter
+- Disallow users to view module and command `help` pages they don't have permission for
+- `playlist` command now defaults to `playlist list` subcommand
+
 ## [0.3.0] - 2020-02-08
 ### Added
 - `queue` subcommands added to modify the current songs in the queue